RTCMultiConnection-v1.3 is ready for tests!
You can test streams renegotiation features; mute/unmute; bandwidth;
direct messages; and much more.

To read what’s new in v1.3 and what you can do:

https://github.com/muaz-khan/WebRTC-Experiment/tree/master/RTCMultiConnection

1. You can send direct messages
2. You can renegotiate streams among all users
3. You can renegotiate streams directly between two unique users
4. You can mute/unmute any stream
5. You can set bandwidth for audio/video streams

The only issue in the moment is “screen-sharing in one-way” is not
working. This issue occurs only with renegotiated streams.

When someone Mute or UnMute stream; currently unable to alert him about
the action; so he can show a “poster” in the place of “video” box. It
will be fixed in the next commits.

An advance demo will be coming soon (in a few days); till that you can
try this “simple” demo:

https://webrtc-experiment.appspot.com/RTCMultiConnection-v1.3/